backend/whisper/requirements.txt
av==12.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.4.0
faster-whisper==1.0.3
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.9.0
grpcio==1.67.0
grpcio-health-checking==1.67.0
grpcio-tools==1.66.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.1.2
onnxruntime==1.19.2
packaging==24.1
protobuf==5.27.2
PyAudio==0.2.14
pydub==0.25.1
pyreadline3==3.5.4
PyYAML==6.0.2
requests==2.32.3
sympy==1.13.3
tokenizers==0.20.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
pyannote.audio
